# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateDocumentRequest {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
}

export interface DocumentTab {
  tabId: string;
  title: string;
}

export interface DocumentInfo {
  documentId: string;
  title: string;
  url: string;
  tabs: DocumentTab[];
}

export interface SubmitNoteRequest {
  documentId: string;
  /**
     * @minLength 1
     * @maxLength 20000
     */
  text: string;
}

export interface NoteSubmissionResult {
  documentId: string;
  tabId: string;
  tabTitle: string;
  createdTab: boolean;
  tags: string[];
  message: string;
}

export interface ErrorResponse {
  message: string;
}