# lib/api-zod/src/generated/types/documentInfo.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentTab } from './documentTab';

export interface DocumentInfo {
  documentId: string;
  title: string;
  url: string;
  tabs: DocumentTab[];
}