# lib/api-zod/src/generated/types/noteSubmissionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NoteSubmissionResult {
  documentId: string;
  tabId: string;
  tabTitle: string;
  createdTab: boolean;
  tags: string[];
  message: string;
}