# lib/api-zod/src/generated/types/submitNoteRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitNoteRequest {
  documentId: string;
  /**
     * @minLength 1
     * @maxLength 20000
     */
  text: string;
}